pyproject.toml
[project]
name = "bltools"
version = "0.1.0"
description = "Modern CLI for downloading manuscripts from the British Library."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "xmltodict>=0.13.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "tenacity>=8.0.0",
]

[project.scripts]
bltools = "bltools.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bltools"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "PL", "RUF"]
ignore = []

[tool.mypy]
strict = true
